+++ test/tests/date.js
@@ -2629,6 +2629,54 @@ package('Date', function () {
 
   });
 
+  group('Month traversal issues', function() {
+
+    // If traversing into a new month don't reset the date if the date was also advanced
+
+    dateEqual(run(new Date(2011, 0, 31), 'advance', [{ month: 1 }]), new Date(2011, 1, 28), 'advanced by month will land on last day if the day does not exist');
+    dateEqual(run(new Date(2011, 0, 31), 'advance', [{ month: 1, day: 3 }]), new Date(2011, 2, 3), 'can still advance days after reset');
+    dateEqual(run(new Date(2011, 2, 31), 'rewind', [{ month: 1 }]), new Date(2011, 1, 28), 'rewind by month will land on last day if the day does not exist');
+    dateEqual(run(new Date(2011, 2, 31), 'rewind', [{ month: 1, day: 3 }]), new Date(2011, 1, 25), 'can still rewind days after reset');
+    dateEqual(run(new Date(2011, 0, 31), 'set', [{ month: 1 }]), new Date(2011, 1, 28), 'set does not cause month traversal');
+    dateEqual(run(new Date(2011, 0, 31), 'set', [{ month: 1, day: 3 }]), new Date(2011, 1, 3), 'set with day does not cause month traversal');
+
+
+    var d = run(new Date(2010, 0, 31), 'set', [{ month: 1 }, true]);
+    dateEqual(d, new Date(2010, 1), 'reset dates will not accidentally traverse into a different month');
+
+  });
+
+  group('DST Issues', function() {
+
+    // These tests are meant to be run in a North American timezone with DST (PDT, MDT, etc)
+
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 1, 45), 'addHours', 1, 'Spring DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 2), 'addHours', -1, 'Spring DST | Reverse');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 1, 1, 45), 'addHours', 15, 'Fall DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 2), 'addHours', -15, 'Fall DST | Reverse');
+
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 1, 45), 'addMinutes', 15, 'Spring DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 2), 'addMinutes', -15, 'Spring DST | Reverse');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 1, 1, 45), 'addMinutes', 15, 'Fall DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 2), 'addMinutes', -15, 'Fall DST | Reverse');
+
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 1, 59, 59), 'addSeconds', 10, 'Spring DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 2), 'addSeconds', -10, 'Spring DST | Reverse');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 1, 1, 59, 59), 'addSeconds', 10, 'Fall DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 2), 'addSeconds', -10, 'Fall DST | Reverse');
+
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 1, 59, 59, 950), 'addMilliseconds', 100, 'Spring DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 2, 8, 2), 'addMilliseconds', -100, 'Spring DST | Reverse');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 1, 1, 59, 59, 950), 'addMilliseconds', 100, 'Fall DST | Forward');
+    assertAddUnitIsNumericallyEqual(new Date(2015, 10, 2), 'addMilliseconds', -100, 'Fall DST | Reverse');
+
+    // Catch for DST inequivalencies
+    // FAILS IN DAMASCUS IN XP!
+    var d = run(new Date(2010, 11, 9, 17), 'set', [{ year: 1998, month: 3, day: 3}, true]);
+    equal(d.getHours(), 0, 'handles DST properly');
+
+  });
+
 });
 
 package('Number', function () {